Rename widgets to panels.
